requirements.txt
jbtech-utils-package
psycopg2
